Add mechanism for adjusting the taphold hold time on a per-key basis
This allows a key to have a signed offset from the base value (e.g.
200ms), adjusting tapholds on weaker fingers to be longer and stronger
fingers to be shorter, depending on preference. To support a wider range
of hold delays, 16 bits are now dedicated to a millisecond counter
instead of 8.

This commit also removes the O+P combo (caps lock), as it was misfiring
too often.

+++ src/taphold.h
@@ -18,7 +18,7 @@ typedef struct taphold_data_t {
     uint8_t row;
     uint8_t col;
     uint8_t layer;
-    uint8_t hold_counter;
+    uint16_t hold_counter;
 } taphold_data_t;
 
 typedef struct taphold_state_t {
